requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
gunicorn==20.1.0
requests==2.26.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.31.0